# api/openapi.yaml
openapi: 3.0.3
info:
  title: DevPulse API
  version: 0.1.0
  description: OpenAPI contract for DevPulse services. This file is the single source of truth.
servers:
  - url: http://localhost:8001
    description: Local py-intelligence (docker-compose port binding)
...
paths:
...
  /api/v1/alerts/system:
    post:
      tags: [ Ingestion ]
      operationId: ingestSystemAlert
      summary: Ingest a system alert event
      description: Accepts a system alert from external monitors (e.g., Prometheus, Datadog) and publishes it asynchronously via RabbitMQ for downstream processing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemAlertDto"
      responses:
        "202":
          description: System alert accepted for asynchronous processing
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
components:
  schemas:
...
    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    SystemAlertDto:
      type: object
      required: [ alertId, source, severity, description, timestamp ]
      properties:
        alertId:
          type: string
          minLength: 1
          description: Unique identifier for the alert
          example: "alert-9912"
        source:
          type: string
          minLength: 1
          description: Source of the alert (e.g., Prometheus, Datadog)
          example: "Prometheus"
        severity:
          type: string
          minLength: 1
          description: Severity level of the alert
          example: "CRITICAL"
        description:
          type: string
          minLength: 1
          description: Detailed description of the alert
          example: "CPU usage on node-1 exceeded 90% for 5 minutes"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of when the alert was fired
        metrics:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional key-value metrics associated with the alert
          example:
            cpu_load: 92.5
            memory_usage: "8GB"